requirements.txt
aiofiles==0.7.0
asgiref==3.3.4
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
fastapi==0.65.2
filelock==3.0.12
future==0.18.2
google-search-results==2.4.1
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
Hypercorn==0.11.2
hyperframe==6.0.1
idna==2.10
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
numpy==1.21.1
packaging==21.0
priority==1.3.0
protobuf==3.17.3
pydantic==1.8.2
pymongo==3.11.4
pyparsing==2.4.7
python-crfsuite==0.9.7
python-dateutil==1.5
python-dotenv==0.19.0
python-multipart==0.0.5
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
sacremoses==0.0.45
scipy==1.7.0
sentencepiece==0.1.91
seqeval==1.2.2
six==1.16.0
soupsieve==2.2.1
starlette==0.14.2
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.61.2
transformers>=3.5.1
typing-extensions==3.10.0.0
unicorn==1.0.3
Unidecode==1.2.0
urllib3==1.26.5
uuid==1.30
uvicorn==0.14.0
wsproto==1.0.0
